requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.0
defusedxml==0.6.0
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.1
django-braces==1.14.0
django-cleanup==5.0.0
django-cors-headers==3.4.0
django-filter==2.3.0
django-heroku==0.3.1
django-oauth-toolkit==1.3.2
djangorestframework==3.11.1
gunicorn==20.0.4
idna==2.10
Markdown==3.2.2
oauthlib==3.1.0
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
PyPDF2==1.26.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
